ospool-cm health check: ignore log messages from before container start
Log files are preserved between restarts so the part of the health check
that counts errors in the NegotiatorLog was constantly reporting a high
count based on old log messages.  Only count log lines from after the
container was started.

Because timestamps for log messages are written in locale-specific time format,
it's easier to parse them in Python.  I didn't rewrite the whole script in Python
because correctly writing `tail` in Python is hard.